network: correctly treat unsigned char bytes in read network stream
fixes some checksum and length mismatches when receiving big packages,
where byte wraps to negative numbers.

add hackish method to convert from signed byte to "unsigned" using int.

+++ src/de/saschahlusiak/freebloks/network/NET_CHAT.java
@@ -4,7 +4,7 @@ import java.io.ByteArrayOutputStream;
 
 public class NET_CHAT extends NET_HEADER {
 	public int client; /* int 8 */
-	public int length; /* int 8 */
+	public int length; /* uint 8 */
 	public String text; /* uint8[] */
 
 	public NET_CHAT(String text) {
@@ -21,7 +21,7 @@ public class NET_CHAT extends NET_HEADER {
 	public NET_CHAT(NET_HEADER from) {
 		super(from);
 		client = buffer[0];
-		length = buffer[1];
+		length = unsigned(buffer[1]);
 		
 		while ((buffer[length + 1] == (byte)'\0') || (buffer[length +1] == (byte)'\n') || (buffer[length + 1]== (byte)'\r')) 
 			length--;
